refactor(multisig slice)
