+++ frontend/src/app/(routes)/multisig/page.tsx
@@ -1,12 +1,7 @@
 import React from 'react';
-import Multisig from './Multisig';
 
 const page = () => {
-  return (
-    <div className="page">
-      <Multisig />
-    </div>
-  );
+  return <div className="page">page</div>;
 };
 
 export default page;
