+++ frontend/src/utils/constants.ts
@@ -5,3 +5,8 @@ export const PROPOSAL_STATUS_VOTING_PERIOD = 'PROPOSAL_STATUS_VOTING_PERIOD';
 export const COSMOS_CHAIN_ID = 'cosmoshub-4';
 export const OFFCHAIN_VERIFICATION_MESSAGE =
   'Resolute offchain verification.\n\nSign an offchain verification message to\nprove your ownership to access the multisig page.';
+/**
+ * The base URL for the Multisig API.
+ * @constant
+ */
+export const API_URL = process.env.REACT_APP_API_URI;
